requirements.txt
flask==2.3.3
flask-cors==4.0.0
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
openai-whisper==20231117
numpy==1.24.3
pyttsx3==2.90

